Adds ability to enter new password, validate it, then update the password in the database.
