+++ index.js
@@ -66,8 +66,6 @@ app.use('/logout', require('./routes/logout'));
 app.use('/profile', require('./routes/profile'));
 app.use('/changePassword/begin', require('./routes/changePasswordBegin'));
 app.use('/changePassword/begin/submit', require('./routes/changePasswordEmailSubmit'));
-app.use('/changePassword/question', require('./routes/changePasswordQuestion'));
-app.use('/changePassword/question/submit', require('./routes/changePasswordQuestionSubmit'));
 app.use('/changePassword', require('./routes/changePassword'));
 app.use('/changePassword/submit', require('./routes/changePasswordSubmit'));
 app.use('*', require('./routes/error404'));
