+++ routes/changePasswordEmailSubmit.js
@@ -6,7 +6,7 @@ router.post('/changePassword/begin/submit', (req, res) => {
     // res.render("changePasswordQuestion");
     const email = req.body.email;
     console.log('submitted email');
-    res.redirect('/changePassword/question?email=' + email);
+    res.redirect('/changePassword/?email=' + email);
 });
 
 module.exports = router;
