+++ views/changePassword.ejs
@@ -13,6 +13,16 @@
                             </div>
                         </div>
                         <form class="signin-form" action='/changePassword/submit' method='post'>
+                            <div>
+                                <p>Your secret question:</p>
+                                <p>
+                                    <%= question %>
+                                </p>
+                            </div>
+                            <div class="form-group mb-3">
+                                <label class="label" for="password">Answer</label>
+                                <input class="form-control" autocomplete="off" name='answer' placeholder='Answer'>
+                            </div>
                             <div class="form-group mb-3">
                                 <label class="label" for="password">New password</label>
                                 <input class="form-control" autocomplete="off" name='password' type='password'
@@ -23,8 +33,9 @@
                                 <input class="form-control" autocomplete="off" name='confirmPassword' type='password'
                                     placeholder='Confirm new password'>
                             </div>
+                            <input type="hidden" name="email" value="<%= email %>">
                             <div class="form-group">
-                                <button type="submit" class="form-control btn btn-primary rounded submit px-3">Change Password</button>
+                                <button type="submit" class="form-control btn btn-primary rounded submit px-3">Change password</button>
                             </div>
                         </form>
                         <p class="text-center">Not a member? <a href="/signup">Sign Up</a></p>
@@ -33,4 +44,4 @@
             </div>
         </div>
     </div>
-    <%- include("templates/footer") %>
+<%- include("templates/footer") %>
