+++ views/changePasswordComplete.ejs
@@ -1,0 +1,7 @@
+<%- include("templates/header") %>
+    <div class="container px-5 mt-5">
+        <br>
+        <h3>Password changed successfully.</h3>
+        <a href="/login">Login</a>
+    </div>
+<%- include("templates/footer") %>
