--- views/changePasswordQuestion.ejs
@@ -1,37 +1,0 @@
-<%- include("templates/header") %>
-    <div class="container px-5">
-        <div class="row justify-content-center">
-            <div class="col-md-12 col-lg-10">
-                <div class="wrap d-md-flex pt-5">
-                    <div class="col-md-4">
-                        <img src="#" class="img-fluid" alt="Password image">
-                    </div>
-                    <div class="login-wrap p-4 p-md-5">
-                        <div class="d-flex">
-                            <div class="w-100">
-                                <h3 class="mb-4">Change Password</h3>
-                            </div>
-                        </div>
-                        <form class="signin-form" action='/changePassword/question/submit' method='post'>
-                            <div>
-                                <p>Your secret question:</p>
-                                <p>
-                                    <%= question %>
-                                </p>
-                            </div>
-                            <div class="form-group mb-3">
-                                <label class="label" for="password">Answer</label>
-                                <input class="form-control" autocomplete="off" name='answer' placeholder='Answer'>
-                            </div>
-                            <input type="hidden" name="email" value="<%= email %>">
-                            <div class="form-group">
-                                <button type="submit" class="form-control btn btn-primary rounded submit px-3">Next</button>
-                            </div>
-                        </form>
-                        <p class="text-center">Not a member? <a href="/signup">Sign Up</a></p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-<%- include("templates/footer") %>
